% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{wasm_global}
\alias{wasm_global}
\alias{wasm_global<-}
\title{Read and write an instance's exported globals}
\usage{
wasm_global(instance, name)

wasm_global(instance, name) <- value
}
\arguments{
\item{instance}{A \code{nanowasm_instance}.}

\item{name}{The export name of the global.}

\item{value}{The new value.}
}
\value{
\code{wasm_global()} returns the value. \verb{wasm_global<-} returns the
instance, whose global has been changed in place.
}
\description{
\code{wasm_global()} returns the current value of an exported global, converted
as described in \code{\link[=wasm_instantiate]{wasm_instantiate()}}. \verb{wasm_global<-} sets a mutable one;
setting an immutable global is an error.
}
\examples{
mod <- wasm_module(system.file("extdata", "sum.wasm", package = "nanowasm"))
inst <- wasm_instantiate(mod)
wasm_global(inst, "heap_top")
inst$alloc(64L)
wasm_global(inst, "heap_top")
wasm_global(inst, "heap_top") <- 4096L
inst$alloc(8L)
}
