% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanowasm-package.R
\docType{package}
\name{nanowasm-package}
\alias{nanowasm}
\alias{nanowasm-package}
\title{nanowasm: Run 'WebAssembly' Modules with a Bundled Interpreter}
\description{
Loads 'WebAssembly' modules from raw vectors or files and calls their exported functions from R. Modules run in a bundled interpreter, so no system 'WebAssembly' runtime is needed. R vectors can be copied into and out of a module's linear memory, and R functions can be supplied as imports. Modules are sandboxed: they have no access to files, network or environment beyond the imports they are given, and traps, stack exhaustion, memory growth and timeouts are reported as classed R conditions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pedrobtz.github.io/nanowasm/}
  \item \url{https://github.com/pedrobtz/nanowasm}
  \item Report bugs at \url{https://github.com/pedrobtz/nanowasm/issues}
}

}
\author{
\strong{Maintainer}: Pedro Baltazar \email{pedrobtz@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Pedro Baltazar \email{pedrobtz@gmail.com} [copyright holder]
}

Other contributors:
\itemize{
  \item Takashi Yamamoto (Author of the bundled toywasm interpreter) [copyright holder]
}

}
\keyword{internal}
